Fix for issue #18 : Backup not working
Since android 11+ doesn't allow writing to directories other than Download and Documents, I have written Platform specific code for android to use the SAF. The SAF provides with a uri and also writes the isar backup file to the selected uri.

Since we would need bytes to write using the SAF, I first create a file in the Download directory which is allowed on android devices. Then I get the bytes of this file and create the backup file in the directory user selected, and delete the file from the Download folder. On other platforms, I keep the functionality as it was.

I didn't have an ios device to test features there, but I have tried to keep the ios functionalities from breaking.